+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 under development
 -----------------
 - Add `Context.GetHeader`, a shortcut of http.Request.Header.Get.
+- Add `WrapHandler` to wrap a HTTP handler as a middleware.
 
 v1.7.0 March 30, 2020
 ---------------------